requirements.txt
opencv-python==4.8.1.78
numpy==1.24.4
pillow==10.1.0
tqdm==4.66.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113